refactor(harness): own image payload handling
