+++ docs/internals/architecture/harness/product-cli-lifecycle-boundary.md
@@ -10,7 +10,7 @@ resource/package wording, Work/Method preparation, or output schemas.
 | --- | --- | --- | --- |
 | Repeated sequential turn invocation and non-final failure disposal in `coding.cli.__main__` | `harness.host.product_host.ProductHostLifecycle.run_turns` | Product supplies turn values, invocation callback, and runtime/session disposal candidates | Coding uses the lifecycle helper for prompt, print, and mode runner loops; no duplicate loop remains |
 | TTY detection for injected streams | `harness.host.product_host.stream_is_tty` | None | Coding does not implement its own `isatty` probe |
-| Prompt/stdin/file/image input normalization | `harness.host.prompt_input` | Coding supplies prompt/file arguments and the image-resize choice | Coding keeps argument parsing and prompt policy; shared resolver has no Coding import |
+| Prompt/stdin/file/image input normalization | `harness.host.prompt_input` owns assembly; `harness.tools.workspace.image_payload` owns image MIME, dimensions, limits, and resizing | Coding supplies prompt/file arguments and the image-resize choice | Coding keeps argument parsing and prompt policy; Host and the read tool consume one image-payload policy |
 | Model listing normalization, sorting, query matching, and metadata table | `harness.session.model_selection` | Coding supplies the model getter and selects JSON/TSV output | Coding keeps preferred-model policy and persistence wording |
 | Command descriptor listing projection | `harness.commands.project_command_descriptor` | Coding supplies descriptors and chooses JSON/TSV output | RPC serializers remain separate; no product wire fields are changed |
 | Skill and installed-plugin listing projections | `harness.resources.skills.project_skill_descriptor`, `harness.resources.plugins.project_installed_plugin` | Coding supplies resource/settings discovery and output format | Harness owns only object-shape normalization; enable/disable and wording remain Coding |
