+++ src/loushang/harness/host/prompt_input.py
@@ -13,13 +13,14 @@ from pathlib import Path
 from typing import TextIO
 
 from loushang.ai.types import ImagePart
-from loushang.harness.tools.workspace.path_utils import resolve_tool_path
-from loushang.harness.tools.workspace.read import (
+from loushang.harness.tools.workspace.image_payload import (
     PillowReadImageResizer,
     detect_image_dimensions,
+    detect_supported_image_mime_type,
     format_image_dimension_note,
     image_exceeds_inline_limits,
 )
+from loushang.harness.tools.workspace.path_utils import resolve_tool_path
 
 
 @dataclass(frozen=True)
@@ -84,7 +85,7 @@ def _process_file_args(
         payload = path.read_bytes()
         if not payload:
             continue
-        mime_type = _detect_supported_image_mime_type(path, payload)
+        mime_type = detect_supported_image_mime_type(path, payload)
         if mime_type is not None:
             original_dimensions = detect_image_dimensions(mime_type, payload)
             dimensions = original_dimensions
@@ -132,26 +133,6 @@ def _process_file_args(
             raise RuntimeError(f"Could not read file {path}: {error}") from error
         text_parts.append(f'<file name="{path}">\n{content}\n</file>\n')
     return "".join(text_parts), images
-
-
-def _detect_supported_image_mime_type(path: Path, payload: bytes) -> str | None:
-    suffix = path.suffix.lower()
-    if suffix in {".jpg", ".jpeg"} and payload.startswith(b"\xff\xd8\xff"):
-        return "image/jpeg"
-    if suffix == ".png" and payload.startswith(b"\x89PNG\r\n\x1a\n"):
-        return "image/png"
-    if suffix == ".gif" and (
-        payload.startswith(b"GIF87a") or payload.startswith(b"GIF89a")
-    ):
-        return "image/gif"
-    if (
-        suffix == ".webp"
-        and len(payload) >= 12
-        and payload.startswith(b"RIFF")
-        and payload[8:12] == b"WEBP"
-    ):
-        return "image/webp"
-    return None
 
 
 def _read_stdin_prompt(stdin: TextIO) -> str | None:
